test(cas): harden CAS concurrent tests + archtest float64 + CheckVersionMatch godoc (PR#464 F4/F5/F7/F11)
F4: add service-level concurrent tests (goroutine race, mem path):
  - configwrite.TestConcurrentUpdate_ExactlyOneSucceeds
  - configpublish.TestConcurrentRollback_ExactlyOneSucceeds
  - flagwrite.TestConcurrentToggle_ExactlyOneSucceeds
  Each test uses a stateless concurrentSafeTxRunner (no mutable Calls field)
  to avoid data-race detector false positives; asserts successes==1 &&
  versionConflicts==1 via atomic counters.

F5: runtime/state/cas/protocol.go CheckVersionMatch godoc: add caller contract
  explaining that key-absent vs version-mismatch disambiguation requires a
  caller pre-check (SELECT FOR UPDATE or GetByKey probe).

F7: tools/archtest/cas_request_schema_test.go: Minimum *int → *float64 in both
  body-schema and queryParams structs; assertions updated to >= 1.0.

F11: identitymanage.TestChangePassword_ConcurrentRequests_ExactlyOneSucceeds:
  split conflicts counter into versionConflicts + loginFailures; assert
  versionConflicts==1 (strict CAS path); loginFailures trigger t.Fatalf with
  diagnostic message.

+++ tools/archtest/cas_request_schema_test.go
@@ -89,9 +89,9 @@ func assertExpectedVersionInBodySchema(t *testing.T, root, contractDir string) {
 
 	var schema struct {
 		Properties map[string]struct {
-			Type        string `json:"type"`
-			Minimum     *int   `json:"minimum"`
-			Description string `json:"description"`
+			Type        string   `json:"type"`
+			Minimum     *float64 `json:"minimum"`
+			Description string   `json:"description"`
 		} `json:"properties"`
 		Required []string `json:"required"`
 	}
@@ -105,7 +105,7 @@ func assertExpectedVersionInBodySchema(t *testing.T, root, contractDir string) {
 		"CAS-CONTRACT-EXPECTED-VERSION-SCHEMA-01: %s expectedVersion.type must be integer", path)
 	require.NotNil(t, ev.Minimum,
 		"CAS-CONTRACT-EXPECTED-VERSION-SCHEMA-01: %s expectedVersion.minimum must be set", path)
-	assert.GreaterOrEqual(t, *ev.Minimum, 1,
+	assert.GreaterOrEqual(t, *ev.Minimum, 1.0,
 		"CAS-CONTRACT-EXPECTED-VERSION-SCHEMA-01: %s expectedVersion.minimum must be ≥ 1", path)
 
 	requiredSet := make(map[string]bool, len(schema.Required))
@@ -129,9 +129,9 @@ func assertExpectedVersionInQueryParams(t *testing.T, root, contractDir string) 
 		Endpoints struct {
 			HTTP struct {
 				QueryParams map[string]struct {
-					Type     string `yaml:"type"`
-					Required bool   `yaml:"required"`
-					Minimum  *int   `yaml:"minimum"`
+					Type     string   `yaml:"type"`
+					Required bool     `yaml:"required"`
+					Minimum  *float64 `yaml:"minimum"`
 				} `yaml:"queryParams"`
 			} `yaml:"http"`
 		} `yaml:"endpoints"`
@@ -148,7 +148,7 @@ func assertExpectedVersionInQueryParams(t *testing.T, root, contractDir string) 
 		"CAS-CONTRACT-EXPECTED-VERSION-SCHEMA-01: %s expectedVersion.required must be true", path)
 	require.NotNil(t, ev.Minimum,
 		"CAS-CONTRACT-EXPECTED-VERSION-SCHEMA-01: %s expectedVersion.minimum must be set", path)
-	assert.GreaterOrEqual(t, *ev.Minimum, 1,
+	assert.GreaterOrEqual(t, *ev.Minimum, 1.0,
 		"CAS-CONTRACT-EXPECTED-VERSION-SCHEMA-01: %s expectedVersion.minimum must be ≥ 1", path)
 }
